Refactor report tab components
