+++ packages/frontend/src/sections/tablero/report-tab/index.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
 import { Box } from '@mui/material'
-import { TableContextProvider } from 'src/components/table/context'
 import FilterComponent from './filter-component'
 import ComponentOne from './component-one'
+import ComponentThree from './component-three'
 import { PROJECT_COUNT_BY_STATE } from '../../../mocks/report'
-import ComponentTree from './component-tree'
 
 export default function ReportTab() {
   const newProjects = PROJECT_COUNT_BY_STATE.new
@@ -14,7 +13,6 @@ export default function ReportTab() {
   const totalProjects = PROJECT_COUNT_BY_STATE.count
 
   return (
-  <React.Fragment>
     <Box>
       <FilterComponent areas={[]} />
       <ComponentOne
@@ -24,15 +22,13 @@ export default function ReportTab() {
         canceledProjects={canceledProjects}
         totalProjects={totalProjects}
       />
+      <ComponentThree
+        newProjects={newProjects}
+        inProgressProjects={inProgressProjects}
+        finishedProjects={finishedProjects}
+        canceledProjects={canceledProjects}
+        totalProjects={totalProjects}
+      />
     </Box>
-      <TableContextProvider>
-        <ComponentTree
-         newProjects={newProjects}
-         inProgressProjects={inProgressProjects}
-         finishedProjects={finishedProjects}
-         canceledProjects={canceledProjects}
-         totalProjects={totalProjects}/>
-      </TableContextProvider>
-      </React.Fragment>
   )
 }
